change: basic new opponent implementation
